fix(cohorts): validate found cohort

+++ src/services/cohortService.js
@@ -1,6 +1,6 @@
 import models from '../database/models';
 
-const { cohort } = models;
+const { cohort, user: users, program } = models;
 
 class cohortService {
   static async addCohort(cohortData) {
@@ -16,11 +16,23 @@ class cohortService {
   }
 
   static async deleteOne(where) {
-    const found = await cohort.findOne({ where });
+    const found = await cohort.findOne({
+      where,
+      include: [
+        // { model: users, as: 'users' },
+        { model: program, as: 'programs' },
+      ],
+    });
+    const user = await users.findOne({ where: { cohort: found.id } });
+
+    if (user) return { error: 'Please remove existing users to the cohort' };
+    if (found.programs)
+      return { error: 'Please remove existing programs of the cohort' };
+
     const deleted = await cohort.destroy({ where });
 
     if (!deleted) return null;
-    return found;
+    return { deleted: found };
   }
 
   static async updateOne(where, data) {
